chore: update header files and source files for consistency

+++ include/msg_queue.h
@@ -29,5 +29,5 @@ extern std::queue<sn001_suc_rsp> internal_msg_q;
 extern char** node_list;     // Array of node IDs (strings)
 extern size_t node_count; // Number of nodes in the list
 extern uint8_t ttl;
-extern int hash_cache_size;
+extern size_t hash_cache_size;
 extern uint8_t seq_id;
